[typescript] Adjust typings to refactoring in `styles`.
