+++ src/styles/index.d.ts
@@ -1,8 +1,8 @@
 export { default as MuiThemeProvider } from './MuiThemeProvider';
-export { default as createBreakpoints } from './breakpoints';
-export { default as createMuiTheme } from './theme';
-export { default as createPalette } from './palette';
-export { default as createTypography } from './typography';
+export { default as createBreakpoints } from './createBreakpoints';
+export { default as createMuiTheme, Theme } from './createMuiTheme';
+export { default as createPalette } from './createPalette';
+export { default as createTypography } from './createTypography';
 export { default as withStyles } from './withStyles';
 export { StyleRules, StyleRulesCallback } from './withStyles';
 export { default as withTheme } from './withTheme';
