+++ src/styles/withTheme.d.ts
@@ -1,4 +1,4 @@
-import { Theme } from './theme';
+import { Theme } from './createMuiTheme';
 
 interface withTheme<P = {}, T extends Theme = Theme> {
   (component: React.ComponentType<P & { theme: T }>): React.ComponentClass<P>;
